Create legacyservices.py
script to enumerate legacy services by attempting to login with an M365 user account. 

The scripts 'guesses' possible domains for SharePoint, SharePoint SOAP service and ADFS. 

Legacy services do not support the use of MFA and therefore will be a target for APTs to brute force the passwords of discovered accounts.